deleted __init__, modified blog file in views and reworked singleBlog and singleUser templates
