+++ templates/singleUser.html
@@ -17,12 +17,13 @@
 </div>
 
 
-<h1>{{userData.username}}'s Blogs</h1>
+<h1>{{oneUser.username}}'s Blogs</h1>
 
 {% for blog in blogData: %}
 <div class="blogContainer">
    <h2>{{blog.title}}</h1>
    <p>{{blog.body}}</p>
+   <p>Posted by: {{oneUser.username}}</p>
 </div>
 {% endfor %}
